backend/requirements.txt
# PetPal AI Backend Dependencies

fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
google-generativeai>=0.8.0
google-genai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
